hwc3: set setInheritRt on Composer and ComposerClient
In Composer3, the binder interface was changed from hwbinder to libinder
due to the interface change from HIDL to AIDL. Unlike hwbinder, inherit
real-time priority is not a default behavior in libbinder. The patch
overrides the createBinder() for IComposer and IComposerClient to set
InheritRt properly when their binders are created.

Bug: 211720161
Test: Check the real-time priority on composer service.

Change-Id: Ie2ead837b1157284d7764d83878e895a37d6c330